move ui components to separate dir
