+++ packages/jazz-inspector/src/app.tsx
@@ -4,7 +4,10 @@ import "./twind.js";
 export { JazzInspector } from "./viewer/new-app.js";
 export { PageStack } from "./viewer/page-stack.js";
 export { Breadcrumbs } from "./viewer/breadcrumbs.js";
-export { Button } from "./viewer/button.js";
-export { Input } from "./viewer/input.js";
+
+export { Button } from "./ui/button.js";
+export { Input } from "./ui/input.js";
+export { Select } from "./ui/select.js";
+export { Icon } from "./ui/icon.js";
 
 export type { PageInfo } from "./viewer/types.js";
