+++ packages/jazz-inspector/src/ui/icon.tsx
