+++ packages/jazz-inspector/src/viewer/value-renderer.tsx
@@ -2,7 +2,7 @@ import clsx from "clsx";
 import { CoID, JsonValue, LocalNode, RawCoValue } from "cojson";
 import React, { useEffect, useState } from "react";
 import { LinkIcon } from "../link-icon.js";
-import { Button } from "./button.js";
+import { Button } from "../ui/button.js";
 import {
   isBrowserImage,
   resolveCoValue,
